requirements.txt
# Core AI/ML Libraries
numpy>=1.24.0
pandas>=2.0.0

# LLM and GenAI Libraries
openai>=1.0.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Agent Frameworks
autogen>=0.2.0

# Vector Databases and Embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3
